pyproject.toml
[project]
name = "math"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "manim>=0.20.1",
    "mutagen>=1.47.0",
    "openai>=2.26.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.2",
]
